Move independent Iceberg helpers to root-safe module
